Hi Guys, I have create the GUI necessary for the data input. I created couple of forms and menu items: 1. Login Window 2. A number of Menu items link to each respective form 3.Mail Delivery Form 4.Route Modification Form 5.Transport Cost Update Form 6. Manager Window Form
To login to the system, run the LogonWindow class, and supply the
username as "test" and password as "12345".
Upon successful login, the main MDI form will show with the menu items.
All other other windows will be child windows.

I will also work on other forms as well for checking expediture,revenu
etc.